v0/backend/requirements.txt
fastapi
uvicorn[standard]
prisma
google-generativeai
pydantic
python-dotenv
nanoid
qrcode[pil]
